logic for capacity seems good.

+++ src/Interfaces/IPlantable.cs
@@ -6,7 +6,7 @@ namespace Trestlebridge.Interfaces
         void AddResource(Sunflower sunflower);
         public int numOfPlants();
         public string shortId();
-        public double Capacity {get;}
+        public int Capacity {get;}
         public string Type { get; set; }
 
     }
